requirements.txt
quart
quart-cors
Waitress
GitPython
pydriller